Remove redundant code
